% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete_info.R
\name{athlete_info}
\alias{athlete_info}
\title{Title}
\usage{
athlete_info(athlete_inp)
}
\arguments{
\item{athlete_inp}{: input athlete name}
}
\value{
returns data set and graphs given the function
}
\description{
Title
}
\examples{
athlete_info("Nesar Ahmad Bahawi")


}
