% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bordering_countries.R
\name{bordering_countries}
\alias{bordering_countries}
\title{bordering_countries Function}
\usage{
bordering_countries(user_country_name)
}
\arguments{
\item{user_country_name}{: A country of your choosing from the bordering_countries data set (contains all countries)}
}
\value{
a list of countries that border user_country_name and a data set that the user can save
}
\description{
A function that takes in a country name in quotes ("") and prints a list of the countries that border it as well as a data set
}
\examples{

library(genworld)
bordering_countries("Albania")

library(genworld)
bordering_countries("China")


}
