% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{carbon_emission}
\alias{carbon_emission}
\title{World Carbon Emissions Dataset}
\format{
\subsection{\code{carbon_emission}}{

A data frame with 2,203 rows and 7 columns:
\describe{
\item{region}{country region number}
\item{country}{name of country}
\item{year}{year}
\item{series}{series}
\item{value}{number of Co2}
\item{footnotes}{notes on data}
\item{source}{source of data}
...
}
}
}
\source{
\url{https://data.un.org/}
}
\usage{
carbon_emission
}
\description{
Dataset consisting information of each country's carbon emissions from 1975-2019
}
\keyword{datasets}
