% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_timeseries.R
\name{carbon_timeseries}
\alias{carbon_timeseries}
\title{Title}
\usage{
carbon_timeseries(countryinp)
}
\arguments{
\item{countryinp}{: A country of your choosing from the carbon_emission data set and has to be in quotes, to see which countries are on the list you can run "unique(carbon_emission$country)"}
}
\value{
a line graph of co2 emissions throughout the years for a specific country of the user's choice
}
\description{
this function takes in a country from the carbon_emission data set and returns a line graph of co2 emissions throughout the years
}
\examples{
library(genworld)
# Say what your function is doing
carbon_timeseries("Albania")

}
