% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries_lgbt_acceptance.R
\name{countries_lgbt_acceptance}
\alias{countries_lgbt_acceptance}
\title{countries_lgbt_acceptance Function}
\usage{
countries_lgbt_acceptance(user_country_name)
}
\arguments{
\item{user_country_name}{: A country of your choosing from the lgbtq_acceptance data set}
}
\value{
The GAI index for user_country_name
}
\description{
A function that takes in a country name in quotes ("") and returns The Global Acceptance Index (GAI) which measures levels of acceptance toward LGBT people and issues.
}
\examples{
countries_lgbt_acceptance("France")
countries_lgbt_acceptance("Albania")


}
