% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{country_boarders}
\alias{country_boarders}
\title{Country Boarders Dataset}
\format{
\subsection{\code{country_boarders}}{

A data frame with 728 rows and 4 columns:
\describe{
\item{country_code}{two letter country code}
\item{country_name}{name of country}
\item{country_boarder_code}{two letter country code that boarders the country in the first column}
\item{country_boarder_name}{name of country that boarders the country in the first column}
...
}
}
}
\source{
\url{https://github.com/geodatasource/country-borders/blob/master/GEODATASOURCE-COUNTRY-BORDERS.CSV}
}
\usage{
country_boarders
}
\description{
A data set containing information about countries that boarder eachother
}
\keyword{datasets}
