% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{country_borders}
\alias{country_borders}
\title{Country Boarders Dataset}
\format{
\subsection{\code{country_borders}}{

A data frame with 728 rows and 4 columns:
\describe{
\item{country_code}{two letter country code}
\item{country_name}{name of country}
\item{country_border_code}{two letter country code that borders the country in the first column}
\item{country_border_name}{name of country that borders the country in the first column}
...
}
}
}
\source{
\url{https://github.com/geodatasource/country-borders/blob/master/GEODATASOURCE-COUNTRY-BORDERS.CSV}
}
\usage{
country_borders
}
\description{
A data set containing information about countries that boarder each other
}
\keyword{datasets}
