% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_population.R
\name{country_population}
\alias{country_population}
\title{World Population (Between 2000-2022)}
\usage{
country_population(X)
}
\arguments{
\item{X}{is a country/territory that a user inputs of their choice}
}
\value{
a numeric vector that is the population of a country/territory for the years 2000, 2010,2015 and 2022
}
\description{
this function takes in a a country/territory and returns the population of the country for the years 2000, 2010,2015 and 2022
}
\examples{
country_population("Albania")

country_population("Algeria")

country_population("Afghanistan")


}
