% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{covid_hospitalizations}
\alias{covid_hospitalizations}
\title{COVID hospitalizations data set}
\format{
\subsection{\code{covid_hospitalizations}}{

A data frame with 32871 rows and 3 columns:
\describe{
\item{entity}{Name of country}
\item{indicator}{Type of indicator: Daily ICU Occupancy}
\item{daily_icu_occupancy}{Number of COVID patients in ICU Daily}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/shivamb/daily-covid19-hospitalisation-data}
}
\usage{
covid_hospitalizations
}
\description{
Data set containing COVID daily hospitalizations from 2020-2022, separated by each country.
}
\keyword{datasets}
