% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{covid_vaccinations}
\alias{covid_vaccinations}
\title{COVID vaccinations data set}
\format{
\subsection{\code{covid_vaccinations}}{

A data frame with 145221 rows and 3 columns:
\describe{
\item{location}{Country where vaccine was administrated}
\item{date}{Date of vaccine}
\item{daily_vaccinations}{Number of doses administered daily}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/rsrishav/covid-vaccination-dataset}
}
\usage{
covid_vaccinations
}
\description{
Data set containing COVID daily vaccinations rom 2020-2022, separated by each country.
}
\keyword{datasets}
