% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_vaccinations.R
\name{daily_vaccinations_function}
\alias{daily_vaccinations_function}
\title{Daily COVID vaccination rates (Between 2020-2022)}
\usage{
daily_vaccinations_function(X)
}
\arguments{
\item{X}{is a country/territory that a user inputs of their choice}
}
\value{
a numeric vector that is the population of a country/territory for the years 2000, 2010,2015 and 2022
}
\description{
this function takes in a a country/territory and returns the population of the country for the years 2000, 2010,2015 and 2022
}
\examples{
daily_vaccinations_function("Afghanistan")

daily_vaccinations_function("Afghanistan")

daily_vaccinations_function("Afghanistan")


}
