% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_jewish_killed.R
\name{graph_jewish_killed}
\alias{graph_jewish_killed}
\title{graph_jewish_killed Function}
\usage{
graph_jewish_killed(country_name)
}
\arguments{
\item{country_name}{: A European country of your choosing from the jewish_deaths dataset}
}
\value{
A graph of the Jewish people killed during the Holocaust compared to the Jewish people alive before the Holocaust and the number of Jewish people killed.
}
\description{
A function that takes a European country in quotes ("") and returns a graph of the Jewish people killed during the Holocaust compared to the Jewish people alive before the Holocaust.
}
\examples{
graph_jewish_killed("France")
graph_jewish_killed("Austria")

}
