% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{historical_olympics}
\alias{historical_olympics}
\title{Historial Olympics Dataset: 1896 - 2016}
\format{
\subsection{\code{historical_olympics}}{

A data frame with 2,203 rows and 7 columns:
\describe{
\item{id}{Player ID}
\item{name}{Player Name}
\item{sex}{Player Sex}
\item{age}{Player age}
\item{height}{Player height}
\item{weight}{Player weight}
\item{team}{Player team country}
\item{noc}{Country code}
\item{games}{Summer vs. Winter games}
\item{year}{Olympic year}
\item{season}{Olympic season}
\item{city}{Hosting city}
\item{sport}{Sport}
\item{event}{Olympic event}
\item{medal}{Medal won by player}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/heesoo37/120-years-of-olympic-history-athletes-and-results}
}
\usage{
historical_olympics
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
