% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{jewish_deaths}
\alias{jewish_deaths}
\title{Jewish Deaths Data Set}
\format{
\subsection{\code{jewish_deaths}}{

A data frame with 44 rows and 3 columns:
\describe{
\item{country}{Name of country}
\item{number}{Number of Jewish people}
\item{type}{whether the number in the number column pertains to Jews alive before the war or Jews killed in the Holocaust}
...
}
}
}
\source{
\url{https://www.statista.com/statistics/1070564/jewish-populations-deaths-by-country/}
}
\usage{
jewish_deaths
}
\description{
Data set containing information pertaining to Jewish deaths in Europe from the Holocaust
}
\keyword{datasets}
