% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgbt_acceptance_graph.R
\name{lgbt_acceptance_graph}
\alias{lgbt_acceptance_graph}
\title{lgbt_acceptance_graph Function}
\usage{
lgbt_acceptance_graph(country_name)
}
\arguments{
\item{country_name}{: A country of your choosing from the lgbtq_acceptance data set}
}
\value{
A graph of the GAI's of all the countries in the lgbtq_acceptance data set with country_name highlighted in red and the country's GAI.
}
\description{
A function that takes in a country name and returns a graph of the GAI's of all the countries in the lgbtq_acceptance data set with yours highlighted in red.
}
\examples{
lgbt_acceptance_graph("China")
lgbt_acceptance_graph("France")


}
