% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{lgbtq_acceptance}
\alias{lgbtq_acceptance}
\title{LGBTQ+ Acceptance Dataset}
\format{
\subsection{\code{lgbtq_acceptance}}{

A data frame with 174 rows and 2 columns:
\describe{
\item{country_name}{name of country}
\item{gai}{The Global Acceptance Index of the country taken in the range of 2017-2020}
...
}
}
}
\source{
\url{https://williamsinstitute.law.ucla.edu/projects/gai/}
}
\usage{
lgbtq_acceptance
}
\description{
A data set containing information about the acceptance of LGBTQ+ peoples:
Social attitudes can heavily influence whether LGBT people are accepted or rejected by employers,
family members, teachers, clergy, and society as a whole. The Global Acceptance Index (GAI) uses data
from more than 2,750 surveys conducted over 30 years to measure changing levels of acceptance toward
LGBT people and issues in 175 countries. The GAI provides each nation with a score of 0 to 10 to indicate their level of acceptance.
}
\keyword{datasets}
