% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sport_medal_info.R
\name{sport_medal_info}
\alias{sport_medal_info}
\title{Title}
\usage{
sport_medal_info(team_inp, sport_inp)
}
\arguments{
\item{team_inp}{: input team name}

\item{sport_inp}{: input sport name}
}
\value{
returns data set and graphs given the function
}
\description{
Title
}
\examples{
sport_medal_info("United States", "Basketball")


}
