% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/world_function.R
\name{world_function}
\alias{world_function}
\title{World Population (Between 2000-2022)}
\usage{
world_function(X)
}
\arguments{
\item{X}{is a country/territory that a user inputs of their choice}
}
\value{
a numeric vector that is the population of a country/territory for the years 2000, 2010,2015 and 2022
}
\description{
this function takes in a a country/territory and returns the population of the country for the years 2000, 2010,2015 and 2022
}
\examples{
library(sonic)
world_function("Albania")

world_function("Algeria")

world_function("Afghanistan")


}
