% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{world_happiness}
\alias{world_happiness}
\title{World Happiness data set}
\format{
\subsection{\code{world_happiness}}{

A data frame with 2089 rows and 12 columns:
\describe{
\item{country_name}{Name of country}
\item{year}{Year}
\item{life_ladder}{self rating of life}
\item{log_gdp_per_capita}{Capital Name}
\item{social_support}{Social Support}
\item{healthy_life_expectancy_at_birth}{Life Expectancy at birth}
\item{freedom_to_make_life_choices}{Freedom}
\item{generosity}{Generosity}
\item{perceptions_of_corruption}{existence of corruption}
\item{positive_affect}{Positive effect}
\item{Negative Effect}{Negative effect}
\item{confidence_in_national_government}{Confidence in government}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/whenamancodes/world-happiness-report}
}
\usage{
world_happiness
}
\description{
Data set containing information pertaining to world happiness and statistics on some factors that may influence its ratings
}
\keyword{datasets}
