% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{world_population}
\alias{world_population}
\title{World Population data set}
\format{
\subsection{\code{world_population}}{

A data frame with 234 rows and 17 columns:
\describe{
\item{rank}{rank by population}
\item{cca3}{Country codes}
\item{country_territory}{Name of the country/territory}
\item{capital}{Capital Name}
\item{continent}{Continent Name}
\item{x2022_population}{Population of country in 2022}
\item{x2020_population}{Population of country in 2020}
\item{x2015_population}{Population of country in 2015}
\item{x2010_population}{Population of country in 2010}
\item{x2000_population}{Population of country in 2000}
\item{x1990_population}{Population of country in 1990}
\item{x1980_population}{Population of country in 1980}
\item{x1970_population}{Population of country in 1970}
\item{area_km2}{Area size of country/territory}
\item{density_per_km2}{Population density per square km}
\item{growth_rate}{population growth rate for each country}
\item{world_population_percentage}{the population for each country}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/d893bc6eb4370c2fd7b87bcf41972963b607202a1683f576700c52e6ecd4ab2a?resource=download}
}
\usage{
world_population
}
\description{
Data set containing information related to world population over the years for each country/region
}
\keyword{datasets}
