% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_olympics.R
\name{athlete_info}
\alias{athlete_info}
\title{Title}
\usage{
athlete_info(athlete_inp)
}
\arguments{
\item{athlete_inp:}{input athlete name}

\item{team_inp:}{input team name}
}
\value{
returns data set and graphs given the function
}
\description{
Title
}
\examples{
library(genworld)
# Say what your function is doing
team_sport_medal_info("China")
team_medal_info("Saudi Arabia")


}
