% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naomi_function.R
\name{bordering_countries}
\alias{bordering_countries}
\title{bordering_countries Function}
\usage{
bordering_countries(user_country_name)
}
\arguments{
\item{user_country_name:}{A country of your choosing from the bordering_countries data set (contains all countries)}
}
\value{
a list of countries that border user_country_name
}
\description{
A function that takes in a country name in quotes ("") and prints a list of the countries that border it.
}
\examples{

library(SONIC)
bordering_countries("Albania")

library(SONIC)
bordering_countries("China")


}
