% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{covid_hospitalizations}
\alias{covid_hospitalizations}
\title{COVID hospitalizations data set}
\format{
\subsection{\code{covid_hospitalizations}}{

A data frame with 145221 rows and 3 columns:
\describe{
\item{entity}{Country where vaccine was administrated}
\item{indicator}{Date of vaccine}
\item{value}{Number of doses administered daily}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/rsrishav/covid-vaccination-dataset}
}
\usage{
covid_hospitalizations
}
\description{
Data set containing COVID daily vaccinations from 2020-2022, separated by each country.
}
\keyword{datasets}
