% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_vaccinations_function.R
\name{daily_vaccinations_function}
\alias{daily_vaccinations_function}
\title{Daily COVID vaccination rates (Between 2020-2022)}
\usage{
daily_vaccinations_function(X)
}
\arguments{
\item{X}{is a country/territory that a user inputs of their choice}
}
\value{
a numeric vector that is the number of daily vaccinations a country has administered from 2020-2022
}
\description{
this function takes in a country and outputs the number of daily vaccinations a country has administered from 2020-2022
}
\examples{
daily_vaccinations_function("Algeria")
daily_vaccinations_function("Belgium")

}
