% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_medal_info.R
\name{team_medal_info}
\alias{team_medal_info}
\title{Title}
\usage{
team_medal_info(team_inp)
}
\arguments{
\item{team_inp}{: input team name}
}
\value{
returns data set and graphs given the function
}
\description{
Title
}
\examples{
team_medal_info("United States")

}
